Prototype changes to use semaphore signalling for Unix
Very much in the quick-prototype phase. Demonstrate that
the CTRL-C logic can be alternatively implemented using a
unnamed POSIX semaphore to signal between signal-handler and blocking
thread.
