+++ Cargo.toml
@@ -16,6 +16,7 @@ rust-version = "1.69.0"
 
 [target.'cfg(unix)'.dependencies]
 nix = { version = "0.29", default-features = false, features = ["fs", "signal"]}
+libc = { version = "^0.2" }
 
 [target.'cfg(windows)'.dependencies]
 windows-sys = { version = "0.59", features = ["Win32_Foundation", "Win32_System_Threading", "Win32_Security", "Win32_System_Console"] }
@@ -25,6 +26,7 @@ windows-sys = { version = "0.59", features = ["Win32_Storage_FileSystem", "Win32
 
 [features]
 termination = []
+unix_use_semaphore = []
 
 [[test]]
 harness = false
