+++ src/platform/unix/mod.rs
@@ -7,13 +7,17 @@
 // notice may not be copied, modified, or distributed except
 // according to those terms.
 
-use crate::error::Error as CtrlcError;
-use nix::unistd;
-use std::os::fd::BorrowedFd;
-use std::os::fd::IntoRawFd;
-use std::os::unix::io::RawFd;
+#[cfg(feature = "unix_use_semaphore")]
+mod semaphore;
 
-static mut PIPE: (RawFd, RawFd) = (-1, -1);
+#[cfg(not(feature = "unix_use_semaphore"))]
+mod pipe;
+
+#[cfg(feature = "unix_use_semaphore")]
+pub use self::semaphore::*;
+
+#[cfg(not(feature = "unix_use_semaphore"))]
+pub use self::pipe::*;
 
 /// Platform specific error type
 pub type Error = nix::Error;
@@ -24,61 +28,11 @@ pub type Signal = nix::sys::signal::Signal;
 extern "C" fn os_handler(_: nix::libc::c_int) {
     // Assuming this always succeeds. Can't really handle errors in any meaningful way.
     unsafe {
-        let fd = BorrowedFd::borrow_raw(PIPE.1);
-        let _ = unistd::write(fd, &[0u8]);
+        #[cfg(feature = "unix_use_semaphore")]
+        os_handler_sem();
+        #[cfg(not(feature = "unix_use_semaphore"))]
+        os_handler_pipe();
     }
-}
-
-// pipe2(2) is not available on macOS, iOS, AIX or Haiku, so we need to use pipe(2) and fcntl(2)
-#[inline]
-#[cfg(any(
-    target_os = "ios",
-    target_os = "macos",
-    target_os = "haiku",
-    target_os = "aix",
-    target_os = "nto",
-))]
-fn pipe2(flags: nix::fcntl::OFlag) -> nix::Result<(RawFd, RawFd)> {
-    use nix::fcntl::{fcntl, FcntlArg, FdFlag, OFlag};
-
-    let pipe = unistd::pipe()?;
-    let pipe = (pipe.0.into_raw_fd(), pipe.1.into_raw_fd());
-
-    let mut res = Ok(0);
-
-    if flags.contains(OFlag::O_CLOEXEC) {
-        res = res
-            .and_then(|_| fcntl(pipe.0, FcntlArg::F_SETFD(FdFlag::FD_CLOEXEC)))
-            .and_then(|_| fcntl(pipe.1, FcntlArg::F_SETFD(FdFlag::FD_CLOEXEC)));
-    }
-
-    if flags.contains(OFlag::O_NONBLOCK) {
-        res = res
-            .and_then(|_| fcntl(pipe.0, FcntlArg::F_SETFL(OFlag::O_NONBLOCK)))
-            .and_then(|_| fcntl(pipe.1, FcntlArg::F_SETFL(OFlag::O_NONBLOCK)));
-    }
-
-    match res {
-        Ok(_) => Ok(pipe),
-        Err(e) => {
-            let _ = unistd::close(pipe.0);
-            let _ = unistd::close(pipe.1);
-            Err(e)
-        }
-    }
-}
-
-#[inline]
-#[cfg(not(any(
-    target_os = "ios",
-    target_os = "macos",
-    target_os = "haiku",
-    target_os = "aix",
-    target_os = "nto",
-)))]
-fn pipe2(flags: nix::fcntl::OFlag) -> nix::Result<(RawFd, RawFd)> {
-    let pipe = unistd::pipe2(flags)?;
-    Ok((pipe.0.into_raw_fd(), pipe.1.into_raw_fd()))
 }
 
 /// Register os signal handler.
@@ -91,23 +45,20 @@ fn pipe2(flags: nix::fcntl::OFlag) -> nix::Result<(RawFd, RawFd)> {
 ///
 #[inline]
 pub unsafe fn init_os_handler(overwrite: bool) -> Result<(), Error> {
-    use nix::fcntl;
     use nix::sys::signal;
 
-    PIPE = pipe2(fcntl::OFlag::O_CLOEXEC)?;
+    #[cfg(feature = "unix_use_semaphore")]
+    init_sem()?;
+    #[cfg(not(feature = "unix_use_semaphore"))]
+    init_pipe()?;
 
-    let close_pipe = |e: nix::Error| -> Error {
-        // Try to close the pipes. close() should not fail,
-        // but if it does, there isn't much we can do
-        let _ = unistd::close(PIPE.1);
-        let _ = unistd::close(PIPE.0);
+    let cleanup = |e: nix::Error| -> Error {
+        #[cfg(feature = "unix_use_semaphore")]
+        cleanup_sem();
+        #[cfg(not(feature = "unix_use_semaphore"))]
+        cleanup_pipe();
         e
     };
-
-    // Make sure we never block on write in the os handler.
-    if let Err(e) = fcntl::fcntl(PIPE.1, fcntl::FcntlArg::F_SETFL(fcntl::OFlag::O_NONBLOCK)) {
-        return Err(close_pipe(e));
-    }
 
     let handler = signal::SigHandler::Handler(os_handler);
     #[cfg(not(target_os = "nto"))]
@@ -123,11 +74,11 @@ pub unsafe fn init_os_handler(overwrite: bool) -> Result<(), Error> {
 
     let sigint_old = match signal::sigaction(signal::Signal::SIGINT, &new_action) {
         Ok(old) => old,
-        Err(e) => return Err(close_pipe(e)),
+        Err(e) => return Err(cleanup(e)),
     };
     if !overwrite && sigint_old.handler() != signal::SigHandler::SigDfl {
         signal::sigaction(signal::Signal::SIGINT, &sigint_old).unwrap();
-        return Err(close_pipe(nix::Error::EEXIST));
+        return Err(cleanup(nix::Error::EEXIST));
     }
 
     #[cfg(feature = "termination")]
@@ -136,54 +87,27 @@ pub unsafe fn init_os_handler(overwrite: bool) -> Result<(), Error> {
             Ok(old) => old,
             Err(e) => {
                 signal::sigaction(signal::Signal::SIGINT, &sigint_old).unwrap();
-                return Err(close_pipe(e));
+                return Err(cleanup(e));
             }
         };
         if !overwrite && sigterm_old.handler() != signal::SigHandler::SigDfl {
             signal::sigaction(signal::Signal::SIGINT, &sigint_old).unwrap();
             signal::sigaction(signal::Signal::SIGTERM, &sigterm_old).unwrap();
-            return Err(close_pipe(nix::Error::EEXIST));
+            return Err(cleanup(nix::Error::EEXIST));
         }
         let sighup_old = match signal::sigaction(signal::Signal::SIGHUP, &new_action) {
             Ok(old) => old,
             Err(e) => {
                 signal::sigaction(signal::Signal::SIGINT, &sigint_old).unwrap();
                 signal::sigaction(signal::Signal::SIGTERM, &sigterm_old).unwrap();
-                return Err(close_pipe(e));
+                return Err(cleanup(e));
             }
         };
         if !overwrite && sighup_old.handler() != signal::SigHandler::SigDfl {
             signal::sigaction(signal::Signal::SIGINT, &sigint_old).unwrap();
             signal::sigaction(signal::Signal::SIGTERM, &sigterm_old).unwrap();
             signal::sigaction(signal::Signal::SIGHUP, &sighup_old).unwrap();
-            return Err(close_pipe(nix::Error::EEXIST));
-        }
-    }
-
-    Ok(())
-}
-
-/// Blocks until a Ctrl-C signal is received.
-///
-/// Must be called after calling [`init_os_handler()`](fn.init_os_handler.html).
-///
-/// # Errors
-/// Will return an error if a system error occurred.
-///
-#[inline]
-pub unsafe fn block_ctrl_c() -> Result<(), CtrlcError> {
-    use std::io;
-    let mut buf = [0u8];
-
-    // TODO: Can we safely convert the pipe fd into a std::io::Read
-    // with std::os::unix::io::FromRawFd, this would handle EINTR
-    // and everything for us.
-    loop {
-        match unistd::read(PIPE.0, &mut buf[..]) {
-            Ok(1) => break,
-            Ok(_) => return Err(CtrlcError::System(io::ErrorKind::UnexpectedEof.into())),
-            Err(nix::errno::Errno::EINTR) => {}
-            Err(e) => return Err(e.into()),
+            return Err(cleanup(nix::Error::EEXIST));
         }
     }
